Initial commit - add files
Add makefile with deployimage rule which publishes the docker image
to your ecr repo of choice (var in makefile)
Add main.py
Add Dockerfile

At this point the program runs in a docker container and navigates
the website until the GET/OWNED button is shown. Currently it will
read the text from this button but will not act upon it.