Update error logging and fix issue on redeem game
Update error logging by replacing the meaning behind the CURRENT_EL
var. Now this is named CURRENT_STEP and the message is more self
explanatory.

Fix issue when redeeming a game where the purchase button was being
clicked too soon.

Now the purchase button and 'I agree with the condition' dialog
are inside of an await as that would break the flow.